% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.r
\name{statistical-inference}
\alias{statistical-inference}
\alias{pl_indicator}
\alias{pl_indicator_form}
\title{Statistical Inference with Persistence Landscapes}
\usage{
pl_indicator(pl, supports, r = 0)

pl_indicator_form(pl, supports, r = 0, p = 1)
}
\arguments{
\item{pl}{A persistent landscape.}

\item{r}{Non-negative number; the power of the coefficient \eqn{1/k} in the
indicator linear form.}

\item{p}{Positive integer or infinity; the power used to compute an integral.}

\item{center}{Double; where to center the moment.}
}
\value{
A persistence landscape (an object of S4 class
'Rcpp_PersistenceLandscape').
}
\description{
Compute test statistics and conduct null hypothesis tests for
persistence data using persistence landscapes. See Section 3 of Bubenik
(2015).
}
\examples{
# multiply by an indicator function
pd <- as_persistence(data.frame(dim = 1, birth = c(0, 1), death = c(2, 4)))
f <- list(c(.5, 3), c(1.5, 2), c(1.75, 2))
# exact landscape
pl <- landscape(pd, degree = 1, exact = TRUE)
par(mfcol = c(2L, 2L), mar = c(0, 2, 0, 2))
plot(pl, xlim = c(0, 5))
# plf <- pl$indicator(f, 0)
plf <- pl_indicator(pl, f)
plot(plf, xlim = c(0, 5))
# discrete landscape
pl <- landscape(pd, degree = 1, min_x = 0, max_x = 5, by = .2)
plot(pl)
# plf <- pl$indicator(f, 0)
plf <- pl_indicator(pl, f)
plot(plf, xlim = c(0, 5))
par(mfrow = c(1L, 1L), mar = c(5.1, 4.1, 4.1, 2.1))

# benchmark one- and two-step computation of indicator-based linear form
bench::mark(
  pl$indicator_form(f, 0, p = 1),
  pl$indicator(f, 0)$integral(p = 1)
)
}
\seealso{
PersistenceLandscape-methods
}
